requirements.txt
google-generativeai==0.5.1
tikapy==1.21
click==8.1.3
colored==1.4.4
pytesseract==0.3.10
pdf2image==1.16.1